@@ -1,5 +1,5 @@
 {
-  "name": "react-draggable",
+  "name": "@tchernyavskyd/react-draggable",
   "version": "4.4.5",
   "description": "React draggable component",
   "main": "build/cjs/cjs.js",
@@ -39,21 +39,21 @@
   },
   "homepage": "https://github.com/react-grid-layout/react-draggable",
   "devDependencies": {
-    "@babel/cli": "^7.17.6",
-    "@babel/core": "^7.17.9",
-    "@babel/eslint-parser": "^7.17.0",
-    "@babel/plugin-proposal-class-properties": "^7.16.7",
-    "@babel/plugin-transform-flow-comments": "^7.16.7",
-    "@babel/preset-env": "^7.16.11",
-    "@babel/preset-flow": "^7.16.7",
-    "@babel/preset-react": "^7.16.7",
-    "@types/react": "^17.0.19",
-    "@types/react-dom": "^17.0.9",
+    "@babel/cli": "^7.22.5",
+    "@babel/core": "^7.22.5",
+    "@babel/eslint-parser": "^7.22.5",
+    "@babel/plugin-proposal-class-properties": "^7.18.6",
+    "@babel/plugin-transform-flow-comments": "^7.22.5",
+    "@babel/preset-env": "^7.22.5",
+    "@babel/preset-flow": "^7.22.5",
+    "@babel/preset-react": "^7.22.5",
+    "@types/react": "^18.2.13",
+    "@types/react-dom": "^18.2.6",
     "assert": "^2.0.0",
-    "babel-loader": "^8.2.5",
-    "babel-plugin-transform-inline-environment-variables": "^0.4.3",
-    "eslint": "^8.14.0",
-    "eslint-plugin-react": "^7.29.4",
+    "babel-loader": "^9.1.2",
+    "babel-plugin-transform-inline-environment-variables": "^0.4.4",
+    "eslint": "^8.43.0",
+    "eslint-plugin-react": "^7.32.2",
     "flow-bin": "^0.176.3",
     "jasmine-core": "^4.1.0",
     "karma": "^6.3.19",
@@ -65,21 +65,21 @@
     "karma-phantomjs-launcher": "^1.0.4",
     "karma-phantomjs-shim": "^1.5.0",
     "karma-webpack": "^5.0.0",
-    "lodash": "^4.17.4",
+    "lodash": "^4.17.21",
     "phantomjs-prebuilt": "^2.1.16",
     "pre-commit": "^1.2.2",
     "process": "^0.11.10",
     "puppeteer": "^13.6.0",
-    "react": "^16.13.1",
-    "react-dom": "^16.13.1",
-    "react-frame-component": "^5.2.1",
-    "react-test-renderer": "^16.13.1",
-    "semver": "^7.3.7",
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0",
+    "react-frame-component": "^5.2.6",
+    "react-test-renderer": "^18.2.0",
+    "semver": "^7.5.2",
     "static-server": "^3.0.0",
-    "typescript": "^4.6.3",
-    "webpack": "^5.72.0",
-    "webpack-cli": "^4.9.2",
-    "webpack-dev-server": "^4.8.1"
+    "typescript": "^5.1.3",
+    "webpack": "^5.87.0",
+    "webpack-cli": "^5.1.4",
+    "webpack-dev-server": "^4.15.1"
   },
   "resolutions": {
     "minimist": "^1.2.5"
@@ -89,11 +89,11 @@
     "test"
   ],
   "dependencies": {
-    "clsx": "^1.1.1",
+    "clsx": "^1.2.1",
     "prop-types": "^15.8.1"
   },
   "peerDependencies": {
-    "react": ">= 16.3.0",
-    "react-dom": ">= 16.3.0"
+    "react": ">= 18.0.0",
+    "react-dom": ">= 18.0.0"
   }
 }
