@@ -3,13 +3,13 @@ import * as React from 'react';
 import PropTypes from 'prop-types';
 import ReactDOM from 'react-dom';
 import clsx from 'clsx';
-import {createCSSTransform, createSVGTransform} from './utils/domFns';
+import {createCSSTransform, createSVGTransform } from './utils/domFns';
 import {canDragX, canDragY, createDraggableData, getBoundPosition} from './utils/positionFns';
 import {dontSetMe} from './utils/shims';
 import DraggableCore from './DraggableCore';
 import type {ControlPosition, PositionOffsetControlPosition, DraggableCoreProps, DraggableCoreDefaultProps} from './DraggableCore';
 import log from './utils/log';
-import type {Bounds, DraggableEventHandler} from './utils/types';
+import type {Bounds, DraggableEventHandler, LastCorePositionChangeHandler, ContentResizeHandler} from './utils/types';
 import type {Element as ReactElement} from 'react';
 
 type DraggableState = {
@@ -19,6 +19,14 @@ type DraggableState = {
   slackX: number, slackY: number,
   isElementSVG: boolean,
   prevPropsPosition: ?ControlPosition,
+  prevContentDimensions: ?{
+    width: number,
+    height: number
+  },
+  lastCorePosition: ?{
+    x: number,
+    y: number
+  }
 };
 
 export type DraggableDefaultProps = {
@@ -44,6 +52,7 @@ export type DraggableProps = {
 //
 
 class Draggable extends React.Component<DraggableProps, DraggableState> {
+  resizeObserver: null|ResizeObserver = null;
 
   static displayName: ?string = 'Draggable';
 
@@ -180,6 +189,7 @@ class Draggable extends React.Component<DraggableProps, DraggableState> {
   // React 16.3+
   // Arity (props, state)
   static getDerivedStateFromProps({position}: DraggableProps, {prevPropsPosition}: DraggableState): ?$Shape<DraggableState> {
+    const newState: $Shape<DraggableState> = {};
     // Set x/y if a new position is provided in props that is different than the previous.
     if (
       position &&
@@ -188,24 +198,20 @@ class Draggable extends React.Component<DraggableProps, DraggableState> {
       )
     ) {
       log('Draggable: getDerivedStateFromProps %j', {position, prevPropsPosition});
-      return {
-        x: position.x,
-        y: position.y,
-        prevPropsPosition: {...position}
-      };
+      newState.x = position.x;
+      newState.y = position.y;
+      newState.prevPropsPosition = {...position};
     }
-    return null;
+    return Object.keys(newState).length ? newState : null;
   }
 
   constructor(props: DraggableProps) {
     super(props);
 
     this.state = {
+      dragged: false,
       // Whether or not we are currently dragging.
       dragging: false,
-
-      // Whether or not we have been dragged before.
-      dragged: false,
 
       // Current transform x and y.
       x: props.position ? props.position.x : props.defaultPosition.x,
@@ -217,7 +223,9 @@ class Draggable extends React.Component<DraggableProps, DraggableState> {
       slackX: 0, slackY: 0,
 
       // Can only determine if SVG after mounting
-      isElementSVG: false
+      isElementSVG: false,
+      prevContentDimensions: null,
+      lastCorePosition: null
     };
 
     if (props.position && !(props.onDrag || props.onStop)) {
@@ -229,14 +237,68 @@ class Draggable extends React.Component<DraggableProps, DraggableState> {
   }
 
   componentDidMount() {
+    const node = this.findDOMNode();
     // Check to see if the element passed is an instanceof SVGElement
-    if(typeof window.SVGElement !== 'undefined' && this.findDOMNode() instanceof window.SVGElement) {
+    if(typeof window.SVGElement !== 'undefined' && node instanceof window.SVGElement) {
       this.setState({isElementSVG: true});
+    }
+
+    // observe draggable content resize
+    if(node){
+      this.resizeObserver = new ResizeObserver((entries) => {
+        entries.forEach(entry => {
+          const { target } = entry;
+          const { width, height } = entry.contentRect;
+          this.onContentResize({width, height, target});
+        });
+      });
+      this.resizeObserver.observe(node);
     }
   }
 
+  onContentResize: ContentResizeHandler = ({ width: newWidth, height: newHeight, target }) => {
+    const prevHeight = this.state.prevContentDimensions?.height;
+    if(this.state.lastCorePosition && prevHeight && newHeight!==prevHeight){
+      const { x: lastCoreX, y: lastCoreY } = this.state.lastCorePosition;
+      const heightDelta = newHeight-prevHeight;
+      const node = this.findDOMNode();
+      if (node) {
+        const handleNode = this.props.handle ? target.querySelector(this.props.handle) || node : node;
+        // get current component position
+        const { left, top } = handleNode.getBoundingClientRect();
+        // Emulate mouse move to the component position
+        const mouseEvent = new MouseEvent('mousemove', {
+          bubbles: true,
+          cancelable: true,
+          clientX: left,
+          clientY: top
+        });
+        // compensate height delta to keep component at the position it was before resize
+        this.onDrag(mouseEvent, {
+          x: lastCoreX,
+          y: top > 0 ? lastCoreY : (heightDelta + top),
+          deltaX: 0,
+          deltaY: top > 0 ? 0 : heightDelta,
+          lastX: lastCoreX,
+          lastY: lastCoreY,
+          node
+        }, true);
+      }
+    }
+    this.setState({prevContentDimensions: {
+        width: newWidth,
+        height: newHeight
+      }});
+  };
+
   componentWillUnmount() {
-    this.setState({dragging: false}); // prevents invariant if unmounted while dragging
+    if (this.state.dragging) {
+      this.setState({dragging: false}); // prevents invariant if unmounted while dragging
+    }
+    // stop observing draggable content resize
+    if (this.resizeObserver) {
+      this.resizeObserver.disconnect();
+    }
   }
 
   // React Strict Mode compatibility: if `nodeRef` is passed, we will use it instead of trying to find
@@ -256,8 +318,8 @@ class Draggable extends React.Component<DraggableProps, DraggableState> {
     this.setState({dragging: true, dragged: true});
   };
 
-  onDrag: DraggableEventHandler = (e, coreData) => {
-    if (!this.state.dragging) return false;
+  onDrag: DraggableEventHandler = (e, coreData, force) => {
+    if (!this.state.dragging && !force) return false;
     log('Draggable: onDrag: %j', coreData);
 
     const uiData = createDraggableData(this, coreData);
@@ -299,6 +361,10 @@ class Draggable extends React.Component<DraggableProps, DraggableState> {
     if (shouldUpdate === false) return false;
 
     this.setState(newState);
+  };
+
+  onLastCorePositionChange: LastCorePositionChangeHandler = (x, y) => {
+    this.setState({lastCorePosition: {x, y}});
   };
 
   onDragStop: DraggableEventHandler = (e, coreData) => {
@@ -383,7 +449,7 @@ class Draggable extends React.Component<DraggableProps, DraggableState> {
     // Reuse the child provided
     // This makes it flexible to use whatever element is wanted (div, ul, etc)
     return (
-      <DraggableCore {...draggableCoreProps} onStart={this.onDragStart} onDrag={this.onDrag} onStop={this.onDragStop}>
+      <DraggableCore {...draggableCoreProps} onStart={this.onDragStart} onDrag={this.onDrag} onStop={this.onDragStop} onLastCorePositionChange={this.onLastCorePositionChange}>
         {React.cloneElement(React.Children.only(children), {
           className: className,
           style: {...children.props.style, ...style},
