@@ -8,7 +8,7 @@ import {createCoreData, getControlPosition, snapToGrid} from './utils/positionFn
 import {dontSetMe} from './utils/shims';
 import log from './utils/log';
 
-import type {EventHandler, MouseTouchEvent} from './utils/types';
+import type {EventHandler, MouseTouchEvent, LastCorePositionChangeHandler} from './utils/types';
 import type {Element as ReactElement} from 'react';
 
 // Simple abstraction for dragging events names.
@@ -56,6 +56,7 @@ export type DraggableCoreDefaultProps = {
   onStop: DraggableEventHandler,
   onMouseDown: (e: MouseEvent) => void,
   scale: number,
+  onLastCorePositionChange: LastCorePositionChangeHandler,
 };
 
 export type DraggableCoreProps = {
@@ -224,6 +225,7 @@ export default class DraggableCore extends React.Component<DraggableCoreProps, D
     onDrag: function(){},
     onStop: function(){},
     onMouseDown: function(){},
+    onLastCorePositionChange: function() {},
     scale: 1,
   };
 
@@ -258,14 +260,23 @@ export default class DraggableCore extends React.Component<DraggableCoreProps, D
       removeEvent(ownerDocument, eventsFor.mouse.stop, this.handleDragStop);
       removeEvent(ownerDocument, eventsFor.touch.stop, this.handleDragStop);
       removeEvent(thisNode, eventsFor.touch.start, this.onTouchStart, {passive: false});
-      if (this.props.enableUserSelectHack) removeUserSelectStyles(ownerDocument);
+      if (this.props.enableUserSelectHack) {
+        // prevent a possible "forced reflow"
+        if (window.requestAnimationFrame) {
+          window.requestAnimationFrame(() => {
+            removeUserSelectStyles(ownerDocument);
+          });
+        } else {
+          removeUserSelectStyles(ownerDocument);
+        }
+      }
     }
   }
 
   // React Strict Mode compatibility: if `nodeRef` is passed, we will use it instead of trying to find
   // the underlying DOM node ourselves. See the README for more information.
   findDOMNode(): ?HTMLElement {
-    return this.props?.nodeRef ? this.props?.nodeRef?.current : ReactDOM.findDOMNode(this);
+    return this.props?.nodeRef?.current ? this.props?.nodeRef?.current : ReactDOM.findDOMNode(this);
   }
 
   handleDragStart: EventHandler<MouseTouchEvent> = (e) => {
@@ -298,7 +309,7 @@ export default class DraggableCore extends React.Component<DraggableCoreProps, D
     // distinguish between individual touches on multitouch screens by identifying which
     // touchpoint was set to this element.
     const touchIdentifier = getTouchIdentifier(e);
-    this.setState({touchIdentifier});
+    this.state.touchIdentifier = touchIdentifier;
 
     // Get the current drag point from the event. This is used as the offset.
     const position = getControlPosition(e, touchIdentifier, this);
@@ -322,12 +333,9 @@ export default class DraggableCore extends React.Component<DraggableCoreProps, D
     // Initiate dragging. Set the current x and y as offsets
     // so we know how much we've moved during the drag. This allows us
     // to drag elements around even if they have been moved, without issue.
-    this.setState({
-      dragging: true,
-
-      lastX: x,
-      lastY: y
-    });
+    this.state.dragging = true;
+    this.state.lastX = x;
+    this.state.lastY = y;
 
     // Add events to the document directly so we catch when the user's mouse/touch moves outside of
     // this element. We use different events depending on whether or not we have detected that this
@@ -337,7 +345,6 @@ export default class DraggableCore extends React.Component<DraggableCoreProps, D
   };
 
   handleDrag: EventHandler<MouseTouchEvent> = (e) => {
-
     // Get the current drag point from the event. This is used as the offset.
     const position = getControlPosition(e, this.state.touchIdentifier, this);
     if (position == null) return;
@@ -372,10 +379,9 @@ export default class DraggableCore extends React.Component<DraggableCoreProps, D
       return;
     }
 
-    this.setState({
-      lastX: x,
-      lastY: y
-    });
+    this.state.lastX = x;
+    this.state.lastY = y;
+    this.props.onLastCorePositionChange(x, y);
   };
 
   handleDragStop: EventHandler<MouseTouchEvent> = (e) => {
@@ -396,30 +402,30 @@ export default class DraggableCore extends React.Component<DraggableCoreProps, D
     const coreEvent = createCoreData(this, x, y);
 
     // Call event handler
-    const shouldContinue = this.props.onStop(e, coreEvent);
-    if (shouldContinue === false || this.mounted === false) return false;
+      const shouldContinue = this.props.onStop(e, coreEvent);
+      if (shouldContinue === false || this.mounted === false) return false;
 
-    const thisNode = this.findDOMNode();
-    if (thisNode) {
-      // Remove user-select hack
-      if (this.props.enableUserSelectHack) removeUserSelectStyles(thisNode.ownerDocument);
-    }
+      const thisNode = this.findDOMNode();
+      if (thisNode) {
+        // Remove user-select hack
+        if (this.props.enableUserSelectHack) removeUserSelectStyles(thisNode.ownerDocument);
+      }
 
-    log('DraggableCore: handleDragStop: %j', coreEvent);
+      log('DraggableCore: handleDragStop: %j', coreEvent);
 
-    // Reset the el.
-    this.setState({
-      dragging: false,
-      lastX: NaN,
-      lastY: NaN
-    });
+      // Reset the el.
+      this.setState({
+        dragging: false,
+        lastX: NaN,
+        lastY: NaN
+      });
 
-    if (thisNode) {
-      // Remove event handlers
-      log('DraggableCore: Removing handlers');
-      removeEvent(thisNode.ownerDocument, dragEventFor.move, this.handleDrag);
-      removeEvent(thisNode.ownerDocument, dragEventFor.stop, this.handleDragStop);
-    }
+      if (thisNode) {
+        // Remove event handlers
+        log('DraggableCore: Removing handlers');
+        removeEvent(thisNode.ownerDocument, dragEventFor.move, this.handleDrag);
+        removeEvent(thisNode.ownerDocument, dragEventFor.stop,  this.handleDragStop);
+      }
   };
 
   onMouseDown: EventHandler<MouseTouchEvent> = (e) => {
@@ -452,7 +458,7 @@ export default class DraggableCore extends React.Component<DraggableCoreProps, D
   render(): React.Element<any> {
     // Reuse the child provided
     // This makes it flexible to use whatever element is wanted (div, ul, etc)
-    return React.cloneElement(React.Children.only(this.props.children), {
+    return React.cloneElement(this.props.children, {
       // Note: mouseMove handler is attached to document so it will still function
       // when the user drags quickly and leaves the bounds of the element.
       onMouseDown: this.onMouseDown,
