@@ -106,10 +106,18 @@ interface EventWithOffset {
   clientX: number, clientY: number
 }
 
+export function getOffsetParentRect(offsetParent: HTMLElement): {
+  left: number;
+  top: number;
+} {
+  const isBody = offsetParent === offsetParent.ownerDocument.body;
+  const { top, left } = isBody ? {left: 0, top: 0} : offsetParent.getBoundingClientRect();
+  return { top, left };
+}
+
 // Get from offsetParent
 export function offsetXYFromParent(evt: EventWithOffset, offsetParent: HTMLElement, scale: number): ControlPosition {
-  const isBody = offsetParent === offsetParent.ownerDocument.body;
-  const offsetParentRect = isBody ? {left: 0, top: 0} : offsetParent.getBoundingClientRect();
+  const offsetParentRect = getOffsetParentRect(offsetParent);
 
   const x = (evt.clientX + offsetParent.scrollLeft - offsetParentRect.left) / scale;
   const y = (evt.clientY + offsetParent.scrollTop - offsetParentRect.top) / scale;
