@@ -1,22 +1,28 @@
 // @flow
-import {isNum, int} from './shims';
-import {getTouch, innerWidth, innerHeight, offsetXYFromParent, outerWidth, outerHeight} from './domFns';
+import { isNum, int } from './shims';
+import { getTouch, innerWidth, innerHeight, offsetXYFromParent, outerWidth, outerHeight } from './domFns';
 
 import type Draggable from '../Draggable';
-import type {Bounds, ControlPosition, DraggableData, MouseTouchEvent} from './types';
+import type { Bounds, ControlPosition, DraggableData, MouseTouchEvent } from './types';
 import type DraggableCore from '../DraggableCore';
 
-export function getBoundPosition(draggable: Draggable, x: number, y: number): [number, number] {
-  // If no bounds, short-circuit and move on
-  if (!draggable.props.bounds) return [x, y];
+export function getBounds(draggable: Draggable): null | {
+  bottom?: number,
+  left?: number,
+  right?: number,
+  top?: number
+} {
+  if (!draggable.props.bounds) {
+    return null;
+  }
 
   // Clone new bounds
-  let {bounds} = draggable.props;
+  let { bounds } = draggable.props;
   bounds = typeof bounds === 'string' ? bounds : cloneBounds(bounds);
   const node = findDOMNode(draggable);
 
   if (typeof bounds === 'string') {
-    const {ownerDocument} = node;
+    const { ownerDocument } = node;
     const ownerWindow = ownerDocument.defaultView;
     let boundNode;
     if (bounds === 'parent') {
@@ -40,6 +46,13 @@ export function getBoundPosition(draggable: Draggable, x: number, y: number): [n
         int(boundNodeStyle.paddingBottom) - int(nodeStyle.marginBottom)
     };
   }
+  return bounds;
+}
+
+export function getBoundPosition(draggable: Draggable, x: number, y: number): [number, number] {
+  const bounds = getBounds(draggable);
+  // If no bounds, short-circuit and move on
+  if (!bounds) return [x, y];
 
   // Keep x and y below right and bottom limits...
   if (isNum(bounds.right)) x = Math.min(x, bounds.right);
@@ -104,6 +117,7 @@ export function createCoreData(draggable: DraggableCore, x: number, y: number): 
 // Create an data exposed by <Draggable>'s events
 export function createDraggableData(draggable: Draggable, coreData: DraggableData): DraggableData {
   const scale = draggable.props.scale;
+
   return {
     node: coreData.node,
     x: draggable.state.x + (coreData.deltaX / scale),
