@@ -1,8 +1,5 @@
 // @flow
 
-// eslint-disable-next-line no-use-before-define
-export type DraggableEventHandler = (e: MouseEvent, data: DraggableData) => void | false;
-
 export type DraggableData = {
   node: HTMLElement,
   x: number, y: number,
@@ -28,3 +25,11 @@ export class TouchEvent2 extends TouchEvent {
 }
 
 export type MouseTouchEvent = MouseEvent & TouchEvent2;
+
+// eslint-disable-next-line no-use-before-define
+export type DraggableEventHandler = (e: MouseEvent, data: DraggableData, force: ?boolean) => void | false;
+
+export type LastCorePositionChangeHandler = (x: number, y: number ) => void | false;
+
+export type ContentResizeHandler = (params: {width: number, height: number, target: Element }) => void | false;
+
